pyproject.toml
[tool.poetry]
name = "epic-events-crm"
version = "0.1.0"
description = ""
authors = ["Damien Vincelot <damien.vincelot@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
psycopg2 = "^2.9.10"
sqlalchemy = "^2.0.37"
passlib = "^1.7.4"
pytest = "^8.3.4"
argon2-cffi = "^23.1.0"
click = "^8.1.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
